+++ hoomd/md/alchemy/__init__.py
@@ -1,0 +1,19 @@
+# Copyright (c) 2009-2022 The Regents of the University of Michigan.
+# Part of HOOMD-blue, released under the BSD 3-Clause License.
+
+"""Alchemical molecular dynamics.
+
+Description of alchmical molecular dynamics.
+
+
+MD updaters (`hoomd.md.update`) perform additional operations during the
+simulation, including rotational diffusion and establishing shear flow.
+Use MD computes (`hoomd.md.compute`) to compute the thermodynamic properties of
+the system state.
+
+See Also:
+    Tutorial: :doc:`tutorial/01-Introducing-Molecular-Dynamics/00-index`
+"""
+
+from . import methods
+from . import pair
