+++ hoomd/md/pair/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(files __init__.py
           pair.py
           aniso.py
-          alch.py
    )
 
 install(FILES ${files}
