+++ hoomd/md/pair/__init__.py
@@ -130,7 +130,6 @@ For anisotropic potentials see `hoomd.md.pair.aniso`
 """
 
 from . import aniso
-from . import alch
 from .pair import (Pair, LJ, Gauss, ExpandedLJ, Yukawa, Ewald, Morse, DPD,
                    DPDConservative, DPDLJ, ForceShiftedLJ, Moliere, ZBL, Mie,
                    ExpandedMie, ReactionField, DLVO, Buckingham, LJ1208, LJ0804,
